requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Werkzeug==3.0.1
email-validator==2.2.0
APScheduler
flask-socketio
python-engineio
python-socketio
pytz
gunicorn
eventlet
psycopg2-binary
python-dotenv
cloudinary>=1.30.0
